Improved development environment
In development the application can now be run to proxy to the react
development server. This resolves the problem where the code needs to be
changed in development to point to the server correcly (which is running
on a different port). I also modified the csproj file to have msbuild
copy the required front-end files to the publish folder when needed.
Therefore the built files no longer need to be copied over and
permanently stored in the wwwroot because in development the requests
are proxied, and in production the files are copied to the publish
folder.