Two additional ways of access in AWS S3 (#277)
* Add Assume Role option

If AWS AssumeRoleArn is specified, the ServiceCollectionExtensions for S3 gets FallbackCredentials (instance profile) then uses those credentials to assume the role specified in AssumeRoleArn. The AssumeRoleCredentials class in AwsIamHelper manages session timeouts automatically by validating and retrieving a new token on use.

* Adding the option for UseInstanceProfile back

As it seems the credentials will not time out. The role assumption works also, and is useful for assuming a role for cross account access.

* Update nuget version of newtonsoft, to pass the build test pipeline